requirements.txt
# OpenSchichtplaner5 CLI Dependencies

# Core dependencies
PyYAML>=6.0
argparse>=1.4.0  # Usually included in Python

# Optional dependencies for enhanced features
colorama>=0.4.6  # Cross-platform colored terminal text
rich>=13.0.0     # Rich text and beautiful formatting (optional)
click>=8.0.0     # Alternative CLI framework (optional)

# Development dependencies (uncomment if needed)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# mypy>=1.0.0